Improvements on automated processes
Auto adds date and value as an note, names are no longer case sensitive,
cleaned some text format implementation (used to save and load).
Some bug fixes, there was a bug where if you open the app for the first time not use it and relaunch it, it won't open up. Also there was a issue with entering debt value. They are fixed.